Fix README Minor UI improvements Set input limitations (multiline, max length)
