+++ src/UI/Screens/Inventory/Inventory/InventoryScreen.styles.ts
@@ -4,9 +4,11 @@ import { LayoutDimensions } from "../../../../Theme/LayoutDimensions";
 
 export const inventoryScreenStyles = StyleSheet.create({
   container: {
-    marginHorizontal: LayoutDimensions.spacing6x,
+    flex: 1,
   },
   listInnerContainer: {
+    paddingHorizontal: LayoutDimensions.spacing6x,
+
     marginTop: LayoutDimensions.spacing6x,
     display: "flex",
   },
